requirements.txt
fastapi
uvicorn[standard]
pydantic-ai
pydantic-ai[gemini]
python-dotenv
